requirements.txt
# Web framework
Flask>=2.2.5

# HTTP requests
requests>=2.28

# HTML parsing
beautifulsoup4>=4.13
lxml>=4.9

# Google search wrapper
googlesearch-python==1.3.0

# Production WSGI server for Flask
gunicorn>=20.1
